+++ src/main/java/wbd/api/client/get/ApiClient_GetCategories.java
@@ -1,15 +1,15 @@
-package wbd.api.client;
+package wbd.api.client.get;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
 
-public class ApiClient_GetFilterOffers {
+public class ApiClient_GetCategories {
 
-    public static Response getFilteredOffers() {
+    public static Response getCategories() {
         return RestAssured
                 .given()
                 .when()
-                .get("/offers/filter?cityName=all&category=all&keyPhrase=all")
+                .get("/categories")
                 .then()
                 .log().all()
                 .extract()
