+++ src/main/java/wbd/api/client/get/ApiClient_GetFilterOffers.java
@@ -1,15 +1,15 @@
-package wbd.api.client;
+package wbd.api.client.get;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
 
-public class ApiClient_GetLocations {
+public class ApiClient_GetFilterOffers {
 
-    public static Response getLocations() {
+    public static Response getFilteredOffers() {
         return RestAssured
                 .given()
                 .when()
-                .get("/locations")
+                .get("/offers/filter?cityName=all&category=all&keyPhrase=all")
                 .then()
                 .log().all()
                 .extract()
